tests/wpt/tests/css/css-sizing/animation/height-no-interpolation.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/4441">
<link rel=help href="https://drafts.csswg.org/css-transitions-2/#transition-property-property">
<link rel=help href="https://drafts.csswg.org/css-sizing/#preferred-size-properties">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
<script>
// height is interpolable, but height:auto is not.
test_no_interpolation({
  property: 'height',
  from: '0px',
  to: 'auto'
});
</script>
